working on order/payment

+++ src/main/java/com/sarabrandserver/product/service/ProductSKUService.java
@@ -1,6 +1,7 @@
 package com.sarabrandserver.product.service;
 
 import com.sarabrandserver.exception.CustomNotFoundException;
+import com.sarabrandserver.exception.ResourceAttachedException;
 import com.sarabrandserver.product.dto.SizeInventoryDTO;
 import com.sarabrandserver.product.entity.ProductDetail;
 import com.sarabrandserver.product.entity.ProductSku;
@@ -33,9 +34,16 @@ public class ProductSKUService {
      * */
     @Transactional
     public void delete(final String sku) {
-        // TODO validate if it contains in order or user session before deleting
+        if (this.productSkuRepo.itemContainsCart(sku) > 0 || this.productSkuRepo.itemBeenBought(sku) > 0) {
+            throw new ResourceAttachedException("cannot delete item as it contains a users cart or order history");
+        }
+
         var obj = productSkuBySKU(sku);
         this.productSkuRepo.delete(obj);
+    }
+
+    public int itemBeenBought(final String sku) {
+        return this.productSkuRepo.itemBeenBought(sku);
     }
 
     public ProductSku productSkuBySKU(String sku) {
